functions/image-recognition/requirements.txt
#python3.11
--extra-index-url https://download.pytorch.org/whl/cpu
Pillow
torch==2.0.1+cpu
torchvision==0.1.6
